Add reversible transaction events (#31)
* delete cli

we no longer try to make a cli in flutter. Doing cli in rust.

* UX changes and new rust code

* linter fixes

new names

* set prefix

* parse new events correctly

running on local

* perfected the query

events where extriniscHash is null are transfers that were the result of a reversible transfer, so we already are listing them.

* asset and formatting fix

* main screen and all transactions screen

* show only reversible on page 1

* shorten addresses

* fix queries, add stop icon

* width fix

* transaction history fixed

* number formatting added

fixed number formatting unit tests for new precision of 9 (previously 12)

* linter fixes

* numberformatter can't handle Decimal directly

* change digits back to 12